+++ src/main/java/com/apicatalog/jsonld/compaction/ValueCompaction.java
@@ -19,9 +19,9 @@ import java.util.Map.Entry;
 import java.util.Optional;
 
 import com.apicatalog.jsonld.JsonLdError;
-import com.apicatalog.jsonld.JsonProvider;
 import com.apicatalog.jsonld.context.ActiveContext;
 import com.apicatalog.jsonld.context.TermDefinition;
+import com.apicatalog.jsonld.json.JsonProvider;
 import com.apicatalog.jsonld.json.JsonUtils;
 import com.apicatalog.jsonld.lang.DirectionType;
 import com.apicatalog.jsonld.lang.Keywords;
