+++ src/main/java/com/apicatalog/jsonld/document/JsonDocument.java
@@ -22,8 +22,8 @@ import java.util.Optional;
 
 import com.apicatalog.jsonld.JsonLdError;
 import com.apicatalog.jsonld.JsonLdErrorCode;
-import com.apicatalog.jsonld.JsonProvider;
 import com.apicatalog.jsonld.http.media.MediaType;
+import com.apicatalog.jsonld.json.JsonProvider;
 import com.apicatalog.jsonld.json.JsonUtils;
 
 import jakarta.json.JsonException;
