+++ src/main/java/com/apicatalog/jsonld/serialization/RdfToObject.java
@@ -20,8 +20,8 @@ import java.io.StringReader;
 import com.apicatalog.jsonld.JsonLdError;
 import com.apicatalog.jsonld.JsonLdErrorCode;
 import com.apicatalog.jsonld.JsonLdOptions.RdfDirection;
+import com.apicatalog.jsonld.json.JsonProvider;
 import com.apicatalog.jsonld.JsonLdVersion;
-import com.apicatalog.jsonld.JsonProvider;
 import com.apicatalog.jsonld.lang.Keywords;
 import com.apicatalog.rdf.RdfLiteral;
 import com.apicatalog.rdf.RdfValue;
